requirements.txt
opencv-python
tflite-support
gspread
google-api-python-client
firebase-admin
protobuf==3.20

